PPF-359 - Not allowed to import from other test files, use helper file instead
